Refactor typedStorage file names.
