+++ src/index.ts
@@ -1,2 +1,2 @@
 export { passThrough } from './serializer/passThrough.js'
-export { typedStorage } from './typedStorage.js'
+export { typedStorage } from './typed-storage.js'
