+++ src/typed-storage.ts
